requirements.txt
python-telegram-bot==20.3
requests==2.32.3
schedule==1.2.1